requirements.txt
python-telegram-bot==20.3
quart==0.18.0
httpx==0.24.1
python-dotenv==1.0.0
aiofiles==23.2.1
